+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.0] - 2026-08-01
+- Added logic to handle git submodules
+- Improved forge handling by utilizing macros from forge-srpm-macros
+- Improved Python detection
+
 ## [0.1.0] - 2026-07-31
 - Initial release v0.1.0
 - Added remote and local repository parsing